+++ examples/comments/get_one_comment.go
@@ -7,22 +7,17 @@ import (
 	"log"
 	"os"
 
-	"github.com/joho/godotenv"
+	_ "github.com/joho/godotenv/autoload"
 	"github.com/k0kubun/pp"
 	itdgo "github.com/slipynil/itd-go"
 )
 
 func main() {
-	godotenv.Load()
-	token := os.Getenv("REFRESH_TOKEN")
-	userAgent := os.Getenv("USER_AGENT")
-
-	cfg := itdgo.Config{
-		RefreshToken: token,
-		UserAgent:    userAgent,
-	}
-
 	ctx := context.Background()
+	cfg := itdgo.Config{
+		RefreshToken: os.Getenv("REFRESH_TOKEN"),
+		UserAgent:    os.Getenv("USER_AGENT"),
+	}
 
 	client, err := itdgo.New(ctx, cfg)
 	if err != nil {
@@ -31,7 +26,7 @@ func main() {
 
 	postID := "c36ae616-765f-4119-8380-5fd8080df2d0"
 
-	commentIterator := client.Comments.NewCommentList(ctx, postID, 100)
+	commentIterator := client.Comments.NewCommentList(postID, 100)
 
 	for commentIterator.HasMore() {
 		comments, err := commentIterator.Next(ctx)