+++ examples/posts/showUserPosts.go
@@ -9,24 +9,16 @@ import (
 	"os"
 	"unicode/utf8"
 
-	"github.com/joho/godotenv"
+	_ "github.com/joho/godotenv/autoload"
 	itdgo "github.com/slipynil/itd-go"
 )
 
 func main() {
-	// Загружаем переменные окружения
-	if err := godotenv.Load(); err != nil {
-		log.Fatal(err)
-	}
-
-	refreshToken := os.Getenv("REFRESH_TOKEN")
-	userAgent := os.Getenv("USER_AGENT")
-
-	cfg := itdgo.Config{
-		RefreshToken: refreshToken,
-		UserAgent:    userAgent,
-	}
 	ctx := context.Background()
+	cfg := itdgo.Config{
+		RefreshToken: os.Getenv("REFRESH_TOKEN"),
+		UserAgent:    os.Getenv("USER_AGENT"),
+	}
 
 	client, err := itdgo.New(ctx, cfg)
 	if err != nil {
@@ -35,7 +27,7 @@ func main() {
 
 	// Создаём итератор для постов пользователя
 	username := "nowkie"
-	iterator := client.Posts.NewUserPosts(ctx, username, 20)
+	iterator := client.Posts.NewUserPosts(username, 20)
 
 	fmt.Printf("Загрузка постов пользователя @%s...\n\n", username)